Add autogeneration of the resume pdf
Implements an endpoint to generate a resume pdf when hit, then redirect
the user to that pdf afterwards. This pdf is cached in the public
folder. On non-development builds, generation of the pdf only occurs if
it does not yet exist. This cache will be cleared on run of `gatsby
build`, `gatsby clean`, or deleting the public folder.